+++ back/src/main/java/com/mongoosereum/dou_survey_zone/v1/api/config/WebSecurityConfig.java
@@ -30,17 +30,19 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         //http 시큐리티 빌더
-        http.cors().and().csrf().disable()
+        http.cors()
+                .and().csrf().disable()
                 .httpBasic().disable() // token을 사용하므로 Basic 인증 disable
                 .sessionManagement() // session 기반이 아님을 명시
-                    .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 .and()
                 .authorizeRequests() // /와 /auth/** 경로는 인증 안해도 됨
                 .antMatchers("/","/api/v1/user/**", "/api/v1/survey/**", "/api/v1/participation/**").permitAll()
                 .anyRequest()
-                .authenticated();
+                .permitAll();
+                //.authenticated();
 
-        //fillter 등록
+        // fillter 등록
         // 매요청 마다
         // CorsFilter 실행한 후에
         // jwtAuthenticationFilter 실행을 한다.
