+++ back/src/main/java/com/mongoosereum/dou_survey_zone/v1/api/controller/MainPageController.java
@@ -11,16 +11,16 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @Slf4j
-@RequestMapping(path = "/api/v1/participation", produces = MediaType.APPLICATION_JSON_VALUE)
-public class ParticipationController {
+@RequestMapping(path = "/api/v1/main", produces = MediaType.APPLICATION_JSON_VALUE)
+public class MainPageController {
 
     @Autowired
-    private PartService service;
+    private PartService partService;
 
     /*누적이용자수*/
-    @GetMapping(path="/accuser")
+    @GetMapping(path="/acc")
     public ResponseEntity<?> accUser() {
-        Part_Acc_MySQL result = service.partACC();
+        Part_Acc_MySQL result = partService.partACC();
         return ResponseEntity.ok().body(result);
     }
