+++ back/src/main/java/com/mongoosereum/dou_survey_zone/v1/api/controller/UserController.java
@@ -95,10 +95,8 @@ public class UserController{
     /* 테스트용 */
     @GetMapping(path="/test")
     public String test(@RequestBody UserDTO userDTO) {
-
         System.out.println(userDTO.getUser_Email());
         System.out.println(userDTO.getUser_Password());
-
         return "success";
     }
