+++ back/src/main/java/com/mongoosereum/dou_survey_zone/v1/api/participation/entity/ACC/ACC_Age.java
@@ -2,22 +2,12 @@ package com.mongoosereum.dou_survey_zone.v1.api.participation.entity.ACC;
 
 import lombok.*;
 
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
+@Data
 public class ACC_Age {
-
     private int age_10;
-
     private int age_20;
-
     private int age_30;
-
     private int age_40;
-
     private int age_50;
-
     private int age_60;
 }
