+++ back/src/main/java/com/mongoosereum/dou_survey_zone/v1/api/participation/entity/ACC/ACC_Gender.java
@@ -2,14 +2,8 @@ package com.mongoosereum.dou_survey_zone.v1.api.participation.entity.ACC;
 
 import lombok.*;
 
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
+@Data
 public class ACC_Gender {
-
     private int man;
-
     private int woman;
 }
