+++ back/src/main/java/com/mongoosereum/dou_survey_zone/v1/api/participation/entity/Part_Acc_MySQL.java
@@ -4,16 +4,10 @@ import com.mongoosereum.dou_survey_zone.v1.api.participation.entity.ACC.ACC_Age;
 import com.mongoosereum.dou_survey_zone.v1.api.participation.entity.ACC.ACC_Gender;
 import lombok.*;
 
-@Getter
-@Setter
+@Data
 @Builder
-@AllArgsConstructor
-@NoArgsConstructor
 public class Part_Acc_MySQL {
-
     private int part_Total;
-
     private ACC_Gender part_Gender;
-
     private ACC_Age part_Age;
 }
